fix: slug based fetchign
