--- pages/[memory_slug].tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import App from '../components/App'
-
-interface Props {}
-
-const Main: React.FC<Props> = ({}) => {
-  return <App />
-}
-
-export default Main
