+++ pages/[slug].tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import MemoryPage from '../components/MemoryPage.tsx'
+import { useRouter } from 'next/router'
+
+const Main = () => {
+  const router = useRouter()
+  const { slug } = router.query
+
+  return <MemoryPage slug={slug as string} />
+}
+
+export default Main
